优化一阶段：内部计算全部迁移到使用int；优化计算过程和数据收集

+++ InfrastSim/TimeDriven/Crafting.cs
@@ -10,7 +10,7 @@ public class Crafting : FacilityBase {
 
     public override bool IsWorking => false;
 
-    public override double MoodConsumeModifier => 0;
+    public override int MoodConsumeModifier => 0;
 
-    public override double EffiencyModifier => 0;
+    public override int EffiencyModifier => 0;
 }
